src/backend/requirements.txt
asyncpg~=0.29.0
fastapi~=0.108.0
uvicorn~=0.25.0
psycopg2-binary~=2.9.5
SQLAlchemy~=2.0.0
alembic~=1.13.1
# dev
pytest~=7.4.0
httpx~=0.26.0
pyjwt~=2.10.1
python-dotenv~=1.0.1